rendered PlaceOfWorshipCards
