+++ main-app/src/Components/PlaceOfWorshipCard.js
@@ -1,11 +1,11 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 
 function PlaceOfWorshipCard(props) {
   const [placeOfWorship, setPlaceOfWorship] = useState({});
 
   return (
     <div>
-      <h2>{props.NAME}</h2>
+      <h2>{props.name}</h2>
     </div>
   )
 };
