+++ main-app/src/Components/Search.js
@@ -1,4 +1,5 @@
 import React from "react";
+import PlacesOfWorshipList from "./PlacesOfWorshipList"
 import Form from "./Form/Form";
 
 function Search() {
@@ -7,6 +8,7 @@ function Search() {
     <div className="search">
       <h1>Search</h1>
       <Form />
+      <PlacesOfWorshipList />
     </div>
   );
 };
